fix: 書籍登録関連のPATHを修正
