+++ api/gateway/native/src/routes/index.ts
@@ -1,4 +1,5 @@
 export { default as common } from './common'
 export { default as v1Auth } from './v1/auth'
 export { default as v1Book } from './v1/book'
+export { default as v1Bookshelf } from './v1/bookshelf'
 export { default as v1User } from './v1/user'
